A modal window template has been added to replenish the user's wallet.

+++ VectorXBackend/Controllers/WebSocketService/UpdateUserDataController.cs
@@ -33,46 +33,57 @@ namespace VectorXBackend.Controllers.WebSocketService
                 var webSocket = await HttpContext.WebSockets.AcceptWebSocketAsync();
                 var buffer = new byte[1024 * 4];
                 var cancellationTokenSource = new CancellationTokenSource();
-                var receiveResult = await webSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
-                // Проверяем, что полученные данные не пустые
-                if (receiveResult.Count > 0)
+                try
                 {
-                    var receivedData = Encoding.UTF8.GetString(buffer, 0, receiveResult.Count);
-                    var userConnectionInfo = JsonSerializer.Deserialize<UserConnectionInfo>(receivedData);
-
-                    //Осуществляем отправку данных пользователю
-                    await SendUserData(userConnectionInfo.UserId, webSocket);
-
-                    //Добавляем сокет пользователя в Web Socket Service, чтобы администратор смог вносить обновления
-                    await _webSocketService.AddOrUpdateSocket(userConnectionInfo, webSocket);
-
-                    // Здесь мы запускаем бесконечный цикл, который ждет входящих сообщений от клиента
-                    while (!cancellationTokenSource.Token.IsCancellationRequested)
+                    var receiveResult = await webSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
+                    // Проверяем, что полученные данные не пустые
+                    if (receiveResult.Count > 0)
                     {
-                        try
-                        {
-                            receiveResult = await webSocket.ReceiveAsync(new ArraySegment<byte>(buffer), cancellationTokenSource.Token);
-                        }
-                        catch (WebSocketException ex)
-                        {
-                            // Обработка исключения, возникающего при попытке чтения из закрытого сокета
-                            // Например, прерывание цикла или другие действия
-                            await _webSocketService.RemoveSocket(userConnectionInfo);
-                            break;
-                        }
+                        var receivedData = Encoding.UTF8.GetString(buffer, 0, receiveResult.Count);
+                        var userConnectionInfo = JsonSerializer.Deserialize<UserConnectionInfo>(receivedData);
 
-                        // Если клиент закрыл соединение, выходим из цикла
-                        if (receiveResult.CloseStatus != null)
+                        if (userConnectionInfo.UserId > 0)
                         {
-                            await _webSocketService.RemoveSocket(userConnectionInfo);
-                            break;
+                            //Осуществляем отправку данных пользователю
+                            await SendUserData(userConnectionInfo.UserId, webSocket);
+
+                            //Добавляем сокет пользователя в Web Socket Service, чтобы администратор смог вносить обновления
+                            await _webSocketService.AddOrUpdateSocket(userConnectionInfo, webSocket);
+
+                            // Здесь мы запускаем бесконечный цикл, который ждет входящих сообщений от клиента
+                            while (!cancellationTokenSource.Token.IsCancellationRequested)
+                            {
+                                try
+                                {
+                                    receiveResult = await webSocket.ReceiveAsync(new ArraySegment<byte>(buffer), cancellationTokenSource.Token);
+                                }
+                                catch (WebSocketException ex)
+                                {
+                                    // Обработка исключения, возникающего при попытке чтения из закрытого сокета
+                                    // Например, прерывание цикла или другие действия
+                                    await _webSocketService.RemoveSocket(userConnectionInfo);
+                                    break;
+                                }
+
+                                // Если клиент закрыл соединение, выходим из цикла
+                                if (receiveResult.CloseStatus != null)
+                                {
+                                    await _webSocketService.RemoveSocket(userConnectionInfo);
+                                    break;
+                                }
+                            }
+
+                            if (receiveResult.CloseStatus != null)
+                            {
+                                await webSocket.CloseAsync(receiveResult.CloseStatus.Value, receiveResult.CloseStatusDescription, CancellationToken.None);
+                            }
                         }
                     }
-
-                    if (receiveResult.CloseStatus != null)
-                    {
-                        await webSocket.CloseAsync(receiveResult.CloseStatus.Value, receiveResult.CloseStatusDescription, CancellationToken.None);
-                    }
+                }
+                catch (WebSocketException ex) when (ex.WebSocketErrorCode == WebSocketError.ConnectionClosedPrematurely)
+                {
+                    // Обработка закрытия соединения WebSocket без завершения close handshake
+                    await webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, "Connection closed prematurely", CancellationToken.None);
                 }
             }
             else
